+++ deploy/assets/js/scroll-flipper.js
@@ -8,7 +8,7 @@
  */
 
 (function() {
-    console.log('[ScrollFlipper v3.143-test] Loading High-Performance Mode...');
+    console.log('[ScrollFlipper v3.144-test] Loading High-Performance Mode...');
 
     let isRunning = false;
     let track, stickyContainer, cards, triggers, cardParent;